% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stump.R
\name{fit_stump}
\alias{fit_stump}
\title{Fit a decision stump}
\usage{
fit_stump(x, y)
}
\arguments{
\item{x}{Numeric vector of features}

\item{y}{Numeric vector of labels}
}
\value{
A list containing stump parameters
}
\description{
Fit a decision stump
}
