% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradboostr_fit.R
\name{gradboostr_fit}
\alias{gradboostr_fit}
\title{Train a GradBoostR model}
\usage{
gradboostr_fit(x, y, method = c("gbm", "rf", "rf_class", "nn"), ...)
}
\arguments{
\item{x}{Matrix or data.frame of features.}

\item{y}{Numeric (regression) or factor/integer (classification).}

\item{method}{One of: "gbm", "rf", "rf_class", "nn".}

\item{...}{Additional hyperparameters passed to the underlying C++ engine.}
}
\value{
A gradboostr_model object.
}
\description{
Train a GradBoostR model
}
