% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradboostr_predict.R
\name{gradboostr_predict}
\alias{gradboostr_predict}
\title{Predict using a GradBoostR model}
\usage{
gradboostr_predict(object, newdata, type = c("response", "prob"), ...)
}
\arguments{
\item{object}{A gradboostr_model from gradboostr_fit().}

\item{newdata}{Matrix or data.frame of features.}

\item{type}{"response" or "prob" (classification only).}

\item{...}{Not used.}
}
\value{
Numeric vector (regression), factor (classification), or probability matrix.
}
\description{
Predict using a GradBoostR model
}
