% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_fit_fast.R
\name{rf_fit_fast}
\alias{rf_fit_fast}
\title{Fast Random Forest Fit}
\usage{
rf_fit_fast(X, y, n_trees = 200, max_depth = 5, mtry = 3)
}
\arguments{
\item{X}{Numeric matrix of predictors}

\item{y}{Numeric vector of responses}

\item{n_trees}{Number of trees}

\item{max_depth}{Maximum depth of each tree}

\item{mtry}{Number of features sampled at each split}
}
\value{
A list representing the forest
}
\description{
Fast Random Forest Fit
}
