% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_predict_fast.R
\name{rf_predict_fast}
\alias{rf_predict_fast}
\title{Fast Random Forest Prediction}
\usage{
rf_predict_fast(X, forest)
}
\arguments{
\item{X}{Numeric matrix of predictors}

\item{forest}{A forest object returned by rf_fit_fast()}
}
\value{
Numeric vector of predictions
}
\description{
Fast Random Forest Prediction
}
